jk: Factor out paramsOption from run.go
